# ImageMongo/src/main/java/com/template/mongodb/controller/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.template.mongodb.controller;

import com.template.mongodb.model.ImageModel;
import com.template.mongodb.model.Product;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-06-04T17:02:42.150979100+05:30[Asia/Calcutta]")

@Validated
@Api(value = "products", description = "the products API")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /products : get all the products
     * It gets all the products from the database
     *
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "get all the products", nickname = "getAllProducts", notes = "It gets all the products from the database", response = Product.class, responseContainer = "List", tags={ "UserProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = Product.class, responseContainer = "List") })
    @RequestMapping(value = "/products",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Product>> getAllProducts() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"image\" : [ { \"imageName\" : \"image\", \"imagedata\" : \"imagedata\", \"imageType\" : \"jpeg\" }, { \"imageName\" : \"image\", \"imagedata\" : \"imagedata\", \"imageType\" : \"jpeg\" } ], \"name\" : \"shoes\", \"category\" : \"hardware\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /products : post a product to database
     * it is used to save the data
     *
     * @param name  (optional)
     * @param category  (optional)
     * @param image  (optional)
     * @return okay accepted (status code 201)
     */
    @ApiOperation(value = "post a product to database", nickname = "saveProduct", notes = "it is used to save the data", response = Product.class, tags={ "UserProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "okay accepted", response = Product.class) })
    @RequestMapping(value = "/products",
        produces = { "application/json" }, 
        consumes = { "multipart/form-data" },
        method = RequestMethod.POST)
    default ResponseEntity<Product> saveProduct(@ApiParam(value = "") @RequestPart(value="name", required=false)  String name,@ApiParam(value = "") @RequestPart(value="category", required=false)  String category,@ApiParam(value = "") @RequestPart(value="image", required=false)  List<ImageModel> image) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"image\" : [ { \"imageName\" : \"image\", \"imagedata\" : \"imagedata\", \"imageType\" : \"jpeg\" }, { \"imageName\" : \"image\", \"imagedata\" : \"imagedata\", \"imageType\" : \"jpeg\" } ], \"name\" : \"shoes\", \"category\" : \"hardware\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}